Final changes (2/2)
